+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 2.5.0-dev
 - raised minimal version of PHP to 8.3
 - deprecated methods BookControl::checkTemplatePath, getPages and setPages
+- added support for PSR-14 event dispatcher
+- deprecated Nette style events
 
 Version 2.4.1
 - re-added support for Latte 3